+++ Sources/TuistGenerator/Generator/LinkGenerator.swift
@@ -112,12 +112,12 @@ final class LinkGenerator: LinkGenerating {
                                  pbxproj: pbxproj,
                                  fileElements: fileElements)
 
-        try generateCopyProductsdBuildPhase(path: path,
-                                            target: target,
-                                            graph: graph,
-                                            pbxTarget: pbxTarget,
-                                            pbxproj: pbxproj,
-                                            fileElements: fileElements)
+        try generateCopyProductsBuildPhase(path: path,
+                                           target: target,
+                                           graph: graph,
+                                           pbxTarget: pbxTarget,
+                                           pbxproj: pbxproj,
+                                           fileElements: fileElements)
 
         try generatePackages(target: target,
                              pbxTarget: pbxTarget,
@@ -361,12 +361,12 @@ final class LinkGenerator: LinkGenerating {
             }
     }
 
-    func generateCopyProductsdBuildPhase(path: AbsolutePath,
-                                         target: Target,
-                                         graph: Graph,
-                                         pbxTarget: PBXTarget,
-                                         pbxproj: PBXProj,
-                                         fileElements: ProjectFileElements) throws
+    func generateCopyProductsBuildPhase(path: AbsolutePath,
+                                        target: Target,
+                                        graph: Graph,
+                                        pbxTarget: PBXTarget,
+                                        pbxproj: PBXProj,
+                                        fileElements: ProjectFileElements) throws
     {
         // If the current target, which is non-shared (e.g., static lib), depends on other focused targets which
         // include Swift code, we must ensure those are treated as dependencies so that Xcode builds the targets