+++ Sources/TuistKit/GraphMappers/ProjectMapperProvider.swift
@@ -3,6 +3,7 @@ import TuistCore
 import TuistGenerator
 import TuistLoader
 import TuistSigning
+import TuistSupport
 
 /// It defines an interface for providing the project mappers to be used for a specific configuration.
 protocol ProjectMapperProviding {
@@ -11,7 +12,7 @@ protocol ProjectMapperProviding {
     func mapper(config: Config) -> ProjectMapping
 }
 
-class ProjectMapperProvider: ProjectMapperProviding {
+final class ProjectMapperProvider: ProjectMapperProviding {
     func mapper(config: Config) -> ProjectMapping {
         var mappers: [ProjectMapping] = []
 
@@ -26,6 +27,13 @@ class ProjectMapperProvider: ProjectMapperProviding {
         // Namespace generator
         if !config.generationOptions.contains(.disableSynthesizedResourceAccessors) {
             mappers.append(SynthesizedResourceInterfaceProjectMapper())
+        }
+
+        // Logfile noise suppression
+        if config.generationOptions.contains(.disableShowEnvironmentVarsInScriptPhases) {
+            mappers.append(
+                TargetProjectMapper(mapper: TargetActionEnvironmentMapper(false))
+            )
         }
 
         // Support for resources in libraries